add fast goid for arm & 386

+++ goid.go
@@ -1,11 +1,11 @@
-// +build amd64 arm64
+// +build amd64 arm64 arm 386
 
 package log
 
-func goid() int64
+func goid() int
 
 // Goid returns the current goroutine id.
 // It exactly matches goroutine id of the stack trace.
 func Goid() int64 {
-	return goid()
+	return int64(goid())
 }
